requirements.txt
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.33.3
h5py==3.7.0
idna==3.3
imageio==2.19.3
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
